feat: 修改获取用户-资产授权的账号列表目录结构;
