+++ apps/perms/api/user_permission/__init__.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 #
-from .common import *
 from .nodes import *
 from .assets import *
 from .nodes_with_assets import *
